Refactor EntityContentOptionForm for destination plugins
