+++ src/Plugin/migrate/destination/Form/DestinationPluginFormBase.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Drupal\feeds_migrate\Plugin\migrate\destination\Form;
+
+use Drupal\feeds_migrate\Plugin\ExternalPluginFormBase;
+
+/**
+ * The base configuration form for migrate destination plugins.
+ */
+abstract class DestinationPluginFormBase extends ExternalPluginFormBase {
+
+
+}
